requirements.txt
flask
flask-cors
pymupdf
easyocr
sentence-transformers
faiss-cpu
transformers
torch
Pillow
numpy